Add workaround for GCC warning
GCC was warning about the expansion of the res parameter to the
localtime_r macro always evaluating to true (because it was taking the
address of an object).

Check the return value of (Windows) localtime_s instead. This changes
the resulting type, but we only use the macro in a boolean context.